feat: add some tests and refactor
