+++ client/src/pages/FormPage/components/FormCheckBox.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useFormContext } from 'react-hook-form';
 
 import { ImCheckboxChecked, ImCheckboxUnchecked } from 'react-icons/im';
-import { ErrorMessage } from '../../../styled/styledComponents';
+import { ErrorMessage } from '../../../styled/smallComponents';
 import { FormData } from './Form';
 
 interface CheckBoxProps {
