+++ client/src/pages/FormPage/components/FormFile.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { useFormContext } from 'react-hook-form';
 import { BiImageAdd } from 'react-icons/bi';
 
-import { ErrorMessage } from '../../../styled/styledComponents';
+import { ErrorMessage } from '../../../styled/smallComponents';
 import { FormData } from './Form';
 
 interface FileProps {
