+++ client/src/styled/SubmitButton.ts
@@ -1,28 +1,5 @@
 import styled from '@emotion/styled';
 
-export const Main = styled.main`
-  flex: 1 1 auto;
-  margin-bottom: 10rem;
-`;
-
-export const ErrorMessage = styled.span`
-  position: absolute;
-  bottom: 0;
-  left: 20px;
-
-  white-space: nowrap;
-  color: #e11d48;
-  animation: toBottom 0.2s forwards;
-`;
-
-export const InputBlock = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-
-  position: relative;
-`;
-
 export const SubmitButton = styled.button`
   height: 40px;
   width: 100%;
